favicon, service, ratings  and list styling
