+++ src/app/app.component.css
@@ -1,7 +1,4 @@
-/* .poo {
-  background-color: aquamarine
-} */
-
 .ahoy-app {
-  font-family: Impact, Haettenschweiler, 'Arial Narrow Bold', sans-serif;
+  font-family: Arial, Helvetica, sans-serif;
+  color: grey;
 }
