+++ src/app/app.component.html
@@ -1,4 +1,3 @@
-<!--The content below is only a placeholder and can be replaced.-->
 <div class="ahoy-app">
   <app-movie></app-movie>
 </div>
