+++ src/app/app.component.ts
@@ -6,7 +6,7 @@ import { Component } from '@angular/core';
   styleUrls: ['./app.component.css']
 })
 export class AppComponent {
-  title = 'Mooveez';
+  title = 'MovieOMDB';
 
   constructor() { }
 }
