+++ src/app/movie/movie.component.html
@@ -1,19 +1,32 @@
-<div class="movie">
+<div class="ahoy-movie">
+  <h1>MovieOMDB</h1>
   <div class="ahoy-search">
-    <input #movieTitle type="text" (keyup.enter)="searchMovie(movieTitle.value)" placeholder="Search for a movie...">
+    <input #movieTitle 
+      type="text"
+      (keyup.enter)="searchMovie(movieTitle.value)"
+      placeholder="Search for a movie...">
     <button (click)="searchMovie(movieTitle.value)">Search</button>
   </div>
 
-  <div *ngIf="isValidSearch()" class="ahoy-movie">
-    <h1>{{ getTitle() }} ({{ getYear() }})</h1>
-    <h2>Directed by {{ getDirector() }}</h2>
-    <h2>Awards: {{ getAwards() }}</h2>
-    <h2>Rating: {{ getRatings() }}<img class="imdb-logo" src="https://cdn1.iconfinder.com/data/icons/square-logo-buttons/512/imdb-black-512.png"></h2>
+  <div *ngIf="isValidSearch()" class="ahoy-movie-valid-result">
+    <h2>{{ getTitle() }} ({{ getYear() }}) - {{ getRuntime() }}</h2>
+    <h3>Directed by {{ getDirector() }}</h3>
+    <h3>Awards: {{ getAwards() }}</h3>
+    <h3>Box Office: {{ getBoxOffice() }}</h3>
+
+    <div class="ahoy-ratings-list">
+      <ul>
+        <li><img class="imdb-logo" src="https://cdn1.iconfinder.com/data/icons/square-logo-buttons/512/imdb-black-512.png">Rating: {{ getImdbRating() }}</li>
+        <li><img class="rt-logo" src="http://www.userlogos.org/files/logos/jumpordie/rottentomatoes_03.png">Rating: {{ getRottenTomatoesRating() }}</li>
+        <li><img class="mc-logo" src="https://upload.wikimedia.org/wikipedia/commons/thumb/2/20/Metacritic.svg/768px-Metacritic.svg.png">Rating: {{ getMetacriticRating() }}</li>
+      </ul>
+    </div>
+
     <img [src]="getImgUrl()">
     <p>{{ getPlot() }}</p>
 
-    <button class="already-watched-button" (click)="addToAlreadyWatched()">Add to Already Watched</button>
-    <button class="watch-list-button" (click)="addToWatchList()">Add to Watch List</button>
+    <button class="ahoy-already-watched-button" (click)="addToAlreadyWatched()">Add to Already Watched</button>
+    <button class="ahoy-watch-list-button" (click)="addToWatchList()">Add to Watch List</button>
   </div>
 
   <div *ngIf="isNotFound()">
@@ -23,21 +36,27 @@
   <div *ngIf="isNotProvided()">
     <p>Please provide a title to search for</p>
   </div>
+</div>
 
-  <div *ngIf="getAlreadyWatchedList().length">
+<div class="ahoy-lists">
+  <div class="ahoy-already-watched" *ngIf="getAlreadyWatchedList().length">
     <h3>Already Watched:</h3>
     <ul>
-      <li *ngFor="let movie of getAlreadyWatchedList(); let i = index">
-        <button (click)="removeFromAlreadyWatched(movie)">Remove</button><button (click)="searchMovie(movie)">Lookup</button>{{ i + 1 }}: {{ movie }}
+      <li *ngFor="let movie of getAlreadyWatchedList()">
+        <a [routerLink]="" (click)="searchMovie(movie)">{{ movie }}</a>
       </li>
     </ul>
   </div>
 
-  <div *ngIf="getWatchList().length">
+  <div class="ahoy-watch-list" *ngIf="getWatchList().length">
     <h3>Watch List:</h3>
     <ul>
       <li *ngFor="let movie of getWatchList(); let i = index">
-        <button (click)="seen(movie)">Seen It!</button> <button (click)="removeFromWatchList(movie)">Remove</button><button (click)="searchMovie(movie)">Lookup</button>{{ i + 1 }}: {{ movie }}
+        <button (click)="seen(movie)" 
+        style="background-color: green; border-radius: 5px; color: white;">Seen It!</button>
+        <button (click)="removeFromWatchList(movie)"
+        style="background-color: red; border-radius: 5px; color: white;">Remove</button>
+        <a [routerLink]="" (click)="searchMovie(movie)">  {{ movie }}</a>
       </li>
     </ul>
   </div>
